+++ packages/f-content-cards/src/services/utils/transformCardData.js
@@ -36,7 +36,8 @@ const transformCardData = card => {
         subtitle = subtitleDescription,
         custom_card_type: type,
         updated,
-        voucher_code: voucherCode
+        voucher_code: voucherCode,
+        headline
     } = extras;
 
     const description = Object.keys(extras)
@@ -58,6 +59,7 @@ const transformCardData = card => {
         description,
         extractedCardId,
         footer,
+        headline,
         icon,
         id,
         image,