+++ packages/f-header/jest.config.js
@@ -15,5 +15,9 @@ module.exports = {
         'jest-serializer-vue'
     ],
 
-    testURL: 'http://localhost/'
+    testURL: 'http://localhost/',
+
+    modulePathIgnorePatterns: [
+        './test/specs/'
+    ]
 };