+++ packages/f-header/test/specs/f-header.component.spec.js
@@ -1,0 +1,12 @@
+import HeaderComponent from '../../test-utils/component-objects/f-header.component';
+
+describe('f-header component tests', () => {
+    it('should display the f-header component', () => {
+        // Arrange
+        browser.url('http://localhost:8080');
+
+        // Assert
+        expect(HeaderComponent.isLogoDisplayed()).toBe(true);
+    });
+});
+