+++ packages/f-registration/test/specs/f-registration.component.spec.js
@@ -1,0 +1,39 @@
+import RegistrationComponent from '../../test-utils/component-objects/f-registration.component';
+
+describe('f-header component tests', () => {
+    it('should display error if first name field is empty', () => {
+
+        // Arrange
+        browser.url('http://localhost:8081');
+
+        const userInfo = {
+            firstName: '',
+            lastName: '',
+            email: '',
+            password: ''
+        };
+
+        // Act
+        RegistrationComponent.submitRegistrationForm(userInfo);
+
+        // Assert
+        expect(RegistrationComponent.isFirstNameEmptyErrorDisplayed()).toBe(true);
+        expect(RegistrationComponent.isLastNameEmptyErrorDisplayed()).toBe(true);
+        expect(RegistrationComponent.isEmailEmptyErrorDisplayed()).toBe(true);
+        expect(RegistrationComponent.isPasswordEmptyErrorDisplayed()).toBe(true);
+    });
+
+    it('should show and be able to click the legal documentation', () => {
+        // Arrange
+        browser.url('http://localhost:8081');
+
+        // Act
+        RegistrationComponent.waitForRegistrationForm();
+
+        // Assert
+        expect(RegistrationComponent.termsAndConditionsLinkCanBeClicked()).toBe(true);
+        expect(RegistrationComponent.privacyPolicyLinkCanBeClicked()).toBe(true);
+        expect(RegistrationComponent.cookiesPolicyLinkCanBeClicked()).toBe(true);
+    });
+});
+