+++ packages/f-services/tests/axios.test.js
@@ -2,9 +2,12 @@ import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
 import axiosServices from '../src/axios';
 
+const { getNetworkDetails, setupResponseTimeRecording } = axiosServices;
+
+let callback;
 let instance;
 let mock;
-let callback;
+let windowSpy;
 
 // eslint-disable-next-line no-unused-vars
 const withDelay = (delay, response) => () => new Promise((resolve, _) => setTimeout(() => resolve(response), delay));
@@ -14,13 +17,74 @@ beforeEach(() => {
     callback = jest.fn();
 });
 
+afterEach(() => {
+    jest.resetAllMocks();
+});
+
+describe('getNetworkDetails', () => {
+    beforeEach(() => {
+        windowSpy = jest.spyOn(window, 'navigator', 'get');
+    });
+
+    it('should return `null` when navigator is undefined', () => {
+        // Arrange
+        windowSpy.mockImplementation(() => undefined);
+
+        // Act
+        const connection = getNetworkDetails();
+
+        // Assert
+        expect(connection).toBe(null);
+    });
+
+    it('should return navigator.connection if it is available', () => {
+        // Arrange
+        windowSpy.mockImplementation(() => ({
+            connection: 'CONNECTION'
+        }));
+
+        // Act
+        const connection = getNetworkDetails();
+
+        // Assert
+        expect(connection).toBe('CONNECTION');
+    });
+
+    it('should return navigator.mozConnection if it is available', () => {
+        // Arrange
+        windowSpy.mockImplementation(() => ({
+            mozConnection: 'MOZ-CONNECTION'
+        }));
+
+        // Act
+        const connection = getNetworkDetails();
+
+        // Assert
+        expect(connection).toBe('MOZ-CONNECTION');
+    });
+
+    it('should return navigator.webkitConnection if it is available', () => {
+        // Arrange
+        windowSpy.mockImplementation(() => ({
+            webkitConnection: 'WEBKIT-CONNECTION'
+        }));
+
+        // Act
+        const connection = getNetworkDetails();
+
+        // Assert
+        expect(connection).toBe('WEBKIT-CONNECTION');
+    });
+});
+
+
 describe('setupResponseTimeRecording', () => {
     it.each([
         'request',
         'response'
     ])('should add %s interceptor', interceptorType => {
         // Arrange
-        axiosServices.setupResponseTimeRecording(instance, callback);
+        setupResponseTimeRecording(instance, callback);
 
         // Assert
         expect(instance.interceptors[interceptorType].handlers.length).toBe(1);
@@ -37,7 +101,7 @@ describe('setupResponseTimeRecording', () => {
         ['response', 'string']
     ])('should not add %s interceptor if callback is %p and not a function', (interceptorType, cb) => {
         // Arrange
-        axiosServices.setupResponseTimeRecording(instance, cb);
+        setupResponseTimeRecording(instance, cb);
 
         // Assert
         expect(instance.interceptors[interceptorType].handlers.length).toBe(0);
@@ -49,7 +113,7 @@ describe('request interceptor', () => {
         // Arrange
         mock = new MockAdapter(instance);
         mock.onGet('/test').reply(200);
-        axiosServices.setupResponseTimeRecording(instance, callback);
+        setupResponseTimeRecording(instance, callback);
 
         // Act
         const response = await instance.get('/test');
@@ -69,7 +133,7 @@ describe('response interceptor', () => {
 
         mock = new MockAdapter(instance);
         mock.onGet('/test').reply(withDelay(delay, [200]));
-        axiosServices.setupResponseTimeRecording(instance, callback);
+        setupResponseTimeRecording(instance, callback);
 
         // Act
         const response = await instance.get('/test');
@@ -85,7 +149,7 @@ describe('response interceptor', () => {
         // Arrange
         mock = new MockAdapter(instance);
         mock.onGet('/test').reply(200);
-        axiosServices.setupResponseTimeRecording(instance, callback);
+        setupResponseTimeRecording(instance, callback);
 
         // Act
         await instance.get('/test');
@@ -98,7 +162,7 @@ describe('response interceptor', () => {
         // Arrange
         mock = new MockAdapter(instance);
         mock.onGet('/test').reply(200);
-        axiosServices.setupResponseTimeRecording(instance, callback);
+        setupResponseTimeRecording(instance, callback);
 
         // Act
         await instance.get('/test');