+++ packages/f-registration/CHANGELOG.md
@@ -3,6 +3,20 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.18.0
+------------------------------
+*September 4, 2020*
+
+### Added
+- Legal text and relevant hyperlinks
+
+v0.17.0
+------------------------------
+*September 1, 2020*
+
+### Added
+- WebDriverIO component tests
+
 v0.16.0
 ------------------------------
 *August 28, 2020*